+++ pojo/src/main/java/com/ybk/dto/match/MatchAScoreDTO.java
@@ -8,10 +8,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class MatchAScoreDTO {
-    private Long matchAId;
-    private Long matchModeId;
+    private Integer matchAId;
+    private Integer matchModeId;
     private String mode; // 比赛模式
     private Integer CurrentGame; // 第几局
-    private Long teamId; // 分数变化的队伍id
+    private Integer teamId; // 分数变化的队伍id
     private Integer plusOrMinus; // 加分或者减分
 }
